알약 사진 변경, 기록 추가 날짜 제한

+++ app/src/main/java/com/teamopendata/mindcareapp/ui/home/listener/OnChangedTaskCompleteListener.java
@@ -1,5 +1,5 @@
 package com.teamopendata.mindcareapp.ui.home.listener;
 
 public interface OnChangedTaskCompleteListener {
-    void onChanged(int position);
+    void onChanged(float position);
 }
